+++ src/lib/filehandler.h
@@ -47,8 +47,8 @@ bool create_tmp_file(const char *filepath);
 FILE *open_tmp_file(sds filepath);
 bool rename_tmp_file(FILE *fp, sds tmp_file, bool write_rc);
 bool write_data_to_file(const char *filepath, const char *data, size_t data_len);
-bool rm_file(sds filepath);
-int try_rm_file(sds filepath);
+bool rm_file(const char *filepath);
+int try_rm_file(const char *filepath);
 
 bool rename_file(const char *src, const char *dst);