+++ src/mympd_api/requests.c
@@ -15,7 +15,7 @@
 #include "src/lib/jsonrpc.h"
 
 /**
- * Pushes a MYMPD_API_CACHES_CREATE to the queue
+ * Pushes a MYMPD_API_CACHES_CREATE event to the queue
  * @return true on success, else false
  */
 bool mympd_api_request_caches_create(void) {
@@ -25,7 +25,7 @@ bool mympd_api_request_caches_create(void) {
 }
 
 /**
- * Pushes a MYMPD_API_JUKEBOX_RESTART to the queue
+ * Pushes a MYMPD_API_JUKEBOX_RESTART event to the queue
  * @param partition partition name
  * @return true on success, else false
  */
@@ -36,20 +36,23 @@ bool mympd_api_request_jukebox_restart(const char *partition) {
 }
 
 /**
- * Pushes a INTERNAL_API_TRIGGER_EVENT_EMIT to the queue
+ * Pushes a INTERNAL_API_TRIGGER_EVENT_EMIT event to the queue
  * @param event trigger event
  * @param partition partition name
+ * @param arguments List of arguments
+ * @param conn_id Mongoose connection id
  * @return true on success, else false
  */
-bool mympd_api_request_trigger_event_emit(enum trigger_events event, const char *partition) {
-    struct t_work_request *request = create_request(REQUEST_TYPE_DISCARD, 0, 0, INTERNAL_API_TRIGGER_EVENT_EMIT, NULL, partition);
-    request->data = tojson_int(request->data, "event", event, false);
-    request->data = sdscatlen(request->data, "}}", 2);
+bool mympd_api_request_trigger_event_emit(enum trigger_events event, const char *partition,
+        struct t_list *arguments, unsigned long conn_id) {
+    struct t_work_request *request = create_request(REQUEST_TYPE_DISCARD, conn_id, 0, INTERNAL_API_TRIGGER_EVENT_EMIT, NULL, partition);
+    struct t_event_data *data = mympd_api_event_data_new(event, arguments);
+    request->extra = data;
     return push_request(request, 0);
 }
 
 /**
- * Pushes a INTERNAL_API_STICKER_FEATURES to the queue
+ * Pushes a INTERNAL_API_STICKER_FEATURES event to the queue
  * @param feat_sticker stickers enabled?
  * @param feat_advsticker advanced sticker support?
  * @return true on success, else false