+++ src/web_server/placeholder.c
@@ -25,7 +25,8 @@ const char *placeholder_image_names[] = {
     [PLACEHOLDER_NA] = "coverimage-notavailable",
     [PLACEHOLDER_PLAYLIST] = "coverimage-playlist",
     [PLACEHOLDER_SMARTPLS] = "coverimage-smartpls",
-    [PLACEHOLDER_STREAM] = "coverimage-stream"
+    [PLACEHOLDER_STREAM] = "coverimage-stream",
+    [PLACEHOLDER_TRANSPARENT] = "coverimage-transparent"
 };
 
 /**