+++ src/web_server/web_server.c
@@ -78,6 +78,7 @@ bool web_server_init(struct mg_mgr *mgr, struct t_config *config, struct t_mg_us
     mg_user_data->placeholder_playlist = sdsempty();
     mg_user_data->placeholder_smartpls = sdsempty();
     mg_user_data->placeholder_folder = sdsempty();
+    mg_user_data->placeholder_transparent = sdsempty();
     sds default_coverimage_names = sdsnew(MYMPD_COVERIMAGE_NAMES);
     mg_user_data->coverimage_names= sds_split_comma_trim(default_coverimage_names, &mg_user_data->coverimage_names_len);
     FREE_SDS(default_coverimage_names);
@@ -350,6 +351,7 @@ static bool parse_internal_message(struct t_work_response *response, struct t_mg
         get_placeholder_image(config->workdir, "coverimage-playlist", &mg_user_data->placeholder_playlist);
         get_placeholder_image(config->workdir, "coverimage-smartpls", &mg_user_data->placeholder_smartpls);
         get_placeholder_image(config->workdir, "coverimage-stream", &mg_user_data->placeholder_stream);
+        get_placeholder_image(config->workdir, "coverimage-transparent", &mg_user_data->placeholder_transparent);
 
         //cleanup
         FREE_SDS(new_mg_user_data->mpd_host);
@@ -843,6 +845,9 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data) {
             else if (mg_match(hm->uri, mg_str("/script/*/*"), NULL)) {
                 script_execute_http(nc, hm, config);
             }
+            else if (mg_match(hm->uri, mg_str("/bgimage/*"), NULL)) {
+                script_execute_bgimage(nc, hm);
+            }
         #endif
             else if (mg_match(hm->uri, mg_str("/assets/coverimage-booklet"), NULL)) {
                 webserver_serve_placeholder_image(nc, hm, mg_user_data->placeholder_booklet);
@@ -864,6 +869,9 @@ static void ev_handler(struct mg_connection *nc, int ev, void *ev_data) {
             }
             else if (mg_match(hm->uri, mg_str("/assets/coverimage-playlist"), NULL)) {
                 webserver_serve_placeholder_image(nc, hm, mg_user_data->placeholder_playlist);
+            }
+            else if (mg_match(hm->uri, mg_str("/assets/coverimage-transparent"), NULL)) {
+                webserver_serve_placeholder_image(nc, hm, mg_user_data->placeholder_transparent);
             }
             else if (mg_match(hm->uri, mg_str("/index.html"), NULL)) {
                 webserver_send_header_redirect(nc, "/", "");